zad1/requirements.txt
# Function dependencies, for example:
google-cloud-bigquery==1.28.1
google-cloud-storage==1.44.0